@@ -30,22 +30,23 @@
     "vuex": "^3.0.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.0.0",
-    "@vue/cli-plugin-eslint": "^3.0.0",
-    "@vue/cli-plugin-unit-mocha": "^3.0.0",
-    "@vue/cli-service": "^3.0.0",
-    "@vue/eslint-config-standard": "^3.0.5",
-    "@vue/test-utils": "^1.0.0-beta.20",
-    "chai": "^4.1.2",
-    "electron": "^2.0.0",
+    "@babel/core": "^7.0.0",
+    "@vue/cli-plugin-babel": "^3.2.0",
+    "@vue/cli-plugin-eslint": "^3.2.1",
+    "@vue/cli-plugin-unit-mocha": "^3.2.0",
+    "@vue/cli-service": "^3.2.0",
+    "@vue/eslint-config-standard": "^4.0.0",
+    "@vue/test-utils": "^1.0.0-beta.27",
+    "chai": "^4.2.0",
+    "electron": "^3.0.11",
     "eslint-plugin-vue": "next",
-    "netlify-lambda": "^1.0.3",
-    "npm-run-all": "^4.1.3",
+    "netlify-lambda": "^1.1.0",
+    "npm-run-all": "^4.1.5",
     "stylus": "^0.54.5",
     "stylus-loader": "^3.0.2",
-    "vue-cli-plugin-electron-builder": "^1.0.0-rc.7",
-    "vue-cli-plugin-vuetify": "^0.4.5",
-    "vue-template-compiler": "^2.5.17",
+    "vue-cli-plugin-electron-builder": "^1.0.0-rc.9",
+    "vue-cli-plugin-vuetify": "^0.4.6",
+    "vue-template-compiler": "^2.5.21",
     "webpack-node-externals": "^1.7.2"
   },
   "main": "background.js"
