@@ -2,5 +2,22 @@ const nodeExternals = require('webpack-node-externals')
 
 module.exports = {
   target: 'node',
-  externals: [nodeExternals()]
+  // optimization: {
+  //   minimize: false
+  // },
+  externals: [nodeExternals()],
+  module: {
+    rules: [
+      {
+        test: /\.js?$/,
+        exclude: /(node_modules|bower_components)/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            babelrc: false
+          }
+        }
+      }
+    ]
+  }
 }
